Refactor GrammarPhrase into its own module
Moved the GrammarPhrase struct and its implementation from phrase_list_grammar.rs to a new grammar_phrase.rs module for better separation of concerns and code organization. Updated imports to use the new module.
